# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClipForge API",
    "description": "AI video/audio clipping platform with B2B API",
    "version": "1.0.0",
    "contact": {
      "name": "ClipForge Support",
      "url": "https://clipforge.dev"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.clipforge.dev",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/v1/auth/oauth/token",
            "scopes": {
              "projects:read": "Read projects",
              "projects:write": "Create/update projects",
              "exports:read": "Read exports",
              "exports:write": "Create exports",
              "admin": "Full access"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Export": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "projectId": { "type": "string" },
          "momentId": { "type": "string" },
          "status": { "type": "string", "enum": ["PENDING", "RENDERING", "COMPLETED", "FAILED"] },
          "format": { "type": "string", "enum": ["MP4", "WEBM", "MOV"] },
          "template": { "type": "string", "nullable": true },
          "artifacts": {
            "type": "object",
            "properties": {
              "mp4_url": { "type": "string" },
              "srt_url": { "type": "string" },
              "thumbnail_url": { "type": "string" }
            }
          },
          "metrics": { "type": "object", "nullable": true },
          "publishedTo": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/projects/{id}/ingest": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Ingest source (file/URL/script)",
        "security": [{ "bearerAuth": [] }, { "apiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceType": { "type": "string", "enum": ["file", "url", "script"] },
                  "sourceUrl": { "type": "string", "nullable": true },
                  "scriptText": { "type": "string", "nullable": true }
                },
                "required": ["sourceType"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingestion started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": { "type": "string" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/clips/{clipId}/export": {
      "post": {
        "tags": ["exports"],
        "summary": "Render clip to MP4/SRT",
        "security": [{ "bearerAuth": [] }, { "apiKeyAuth": [] }],
        "parameters": [
          { "name": "clipId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "format": { "type": "string", "enum": ["MP4", "WEBM", "MOV"], "default": "MP4" },
                  "template": { "type": "string", "nullable": true },
                  "brandKitId": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Export job queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Export" }
              }
            }
          }
        }
      }
    },
    "/v1/exports/{id}": {
      "get": {
        "tags": ["exports"],
        "summary": "Get export status",
        "security": [{ "bearerAuth": [] }, { "apiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Export details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Export" }
              }
            }
          }
        }
      }
    },
...
  }
}